+++ packages/server/src/http/internals/HTTPHandlerOptions.ts
@@ -1,4 +1,3 @@
-import { OnErrorFunction } from '../../internals/OnErrorFunction';
 import {
   AnyRouter,
   inferRouterContext,
@@ -8,6 +7,7 @@ import {
 import { TRPCResponse } from '../../rpc';
 import { TRPCError } from '../../TRPCError';
 import { ResponseMeta } from '../ResponseMeta';
+import { BaseHandlerOptions } from '../../internals/BaseHandlerOptions';
 
 type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
   data: TRPCResponse<unknown, inferRouterError<TRouter>>[];
@@ -19,17 +19,6 @@ type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
   type: ProcedureType | 'unknown';
   errors: TRPCError[];
 }) => ResponseMeta;
-
-/**
- * Base interface for any response handler
- */
-export interface BaseHandlerOptions<TRouter extends AnyRouter, TRequest> {
-  onError?: OnErrorFunction<TRouter, TRequest>;
-  batching?: {
-    enabled: boolean;
-  };
-  router: TRouter;
-}
 
 /**
  * Base interface for anything using HTTP
